chore: implement DepreciablePossession

+++ src/main/java/school/hei/PatrimonyManager/DepreciationRate.java
@@ -6,6 +6,6 @@ import java.time.Duration;
 
 @Data
 public class DepreciationRate {
-    private final int rate;
+    private final Double rate;
     private Duration period;
 }
